% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_intervals.R
\name{make_intervals}
\alias{make_intervals}
\title{Makes a data frame of time intervals based on a input vector of dates}
\usage{
make_intervals(x, labels = null)
}
\arguments{
\item{x}{Input vector of dates. These correspond to the starting dates
of each interval. The interval ends the day prior to the start
of the next interval.}

\item{labels}{A vector of labels that uniquely identifies each interval.}
}
\description{
Makes a data frame of time intervals based on a input vector of dates
}
\examples{
x <- ymd('2020-1-1', '2020-4-1', '2020-7-1', '2020-10-1', '2021-1-1')
make_intervals(x, names=c('Q1','Q2','Q3','Q4'))
}
